backend/pyproject.toml
[project]
name = "finance-backend"
version = "0.1.0"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.34.0",
  "pydantic==2.10.4",
  "pydantic-settings==2.7.0",
  "sqlalchemy==2.0.36",
  "psycopg[binary]==3.2.3",
  "alembic==1.14.0",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.0.1",
  "python-multipart==0.0.20",
  "xlsxwriter==3.2.0",
  "PyJWT>=2.0.0",
]
requires-python = ">=3.11"
